+++ config/vite.vue.dev.config.js
@@ -1,0 +1,13 @@
+import { defineConfig } from 'vite'
+
+import viteReactConfig from './vite.vue.config'
+
+export default defineConfig({
+  ...viteReactConfig,
+  build: {
+    ...viteReactConfig.build,
+    watch: {
+      buildDelay: 500
+    }
+  }
+})