+++ examples/react-apollo/vite.config.js
@@ -1,9 +1,11 @@
 import { defineConfig } from 'vite'
-import tsconfigPaths from 'vite-tsconfig-paths'
 
 import react from '@vitejs/plugin-react'
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [tsconfigPaths(), react()]
+  optimizeDeps: {
+    include: ['react/jsx-runtime']
+  },
+  plugins: [react()]
 })