+++ examples/vue-apollo/vite.config.js
@@ -1,5 +1,4 @@
 import { defineConfig } from 'vite'
-import tsconfigPaths from 'vite-tsconfig-paths'
 
 import vue from '@vitejs/plugin-vue'
 import vuetify from '@vuetify/vite-plugin'
@@ -7,7 +6,6 @@ import vuetify from '@vuetify/vite-plugin'
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [
-    tsconfigPaths(),
     vue(),
     vuetify({
       autoImport: true