+++ packages/hasura-storage-js/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @nhost/hasura-storage-js
 
+## 0.2.1
+
+### Patch Changes
+
+- 584976d: - publishable directory structure changes (ESM, CJS and UMD included in the output)
+  - build system improvements
+  - fixed some bundling concerns (https://github.com/nhost/nhost/issues/428)
+
 ## 0.2.0
 
 ### Minor Changes