+++ packages/nextjs/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @nhost/nextjs
 
+## 1.2.3
+
+### Patch Changes
+
+- 584976d: - publishable directory structure changes (ESM, CJS and UMD included in the output)
+  - build system improvements
+  - fixed some bundling concerns (https://github.com/nhost/nhost/issues/428)
+- Updated dependencies [584976d]
+  - @nhost/core@0.5.3
+  - @nhost/nhost-js@1.1.10
+  - @nhost/react@0.7.3
+
 ## 1.2.2
 
 ### Patch Changes